IndiaBix/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.10.4
argostranslate==1.10.0
beautifulsoup4==4.12.3
blis==1.3.3
bs4==0.0.2
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudscraper==1.2.71
confection==0.1.5
cryptography==46.0.3
ctranslate2==4.6.2
cymem==2.0.13
deep-translator==1.11.4
dill==0.4.0
distro==1.9.0
emoji==2.15.0
exejs==0.0.6
filelock==3.20.1
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-translate==3.23.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jh2==5.0.10
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.6.1
niquests==3.16.1
numpy==2.4.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pathos==0.3.4
pillow==12.0.0
pox==0.3.6
ppft==1.7.7
preshed==3.0.12
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qh3==1.5.6
regex==2025.11.3
reportlab==4.0.7
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sacremoses==0.1.1
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
soupsieve==2.8.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stanza==1.10.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
torch==2.9.1
tqdm==4.67.1
translators==6.0.1
triton==3.5.1
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
urllib3-future==2.15.901
wasabi==1.1.3
wassima==2.0.3
weasel==0.4.3
websockets==15.0.1
wrapt==2.0.1
xx_sent_ud_sm @ https://github.com/explosion/spacy-models/releases/download/xx_sent_ud_sm-3.8.0/xx_sent_ud_sm-3.8.0-py3-none-any.whl#sha256=1b5fb9321723b395f76458d2a13b7a3673c537eb58c0a9df3778f4e0950180b4
